// Presentation/AudioTourApp/reference_ui/src/types.ts
export type LocalizedString = { vi: string; en: string };

export type POI = {
  id: string;
  name: LocalizedString;
  description: LocalizedString;
  lat: number;
  lng: number;
  played: boolean;
  imageUrl?: string;
};

export type Tour = {
  id: string;
  name: LocalizedString;
  description: LocalizedString;
  poiIds: string[];
  imageUrl?: string;
};